+++ Source/Core/DeveloperError.js
@@ -4,6 +4,7 @@ define([
     ], function(
         defined) {
     'use strict';
+    /*global console*/
 
     /**
      * Constructs an exception object that is thrown due to a developer error, e.g., invalid argument,
@@ -51,10 +52,12 @@ define([
          * @readonly
          */
         this.stack = stack;
+
+        console.error(toString(this));
     }
 
     DeveloperError.prototype.toString = function() {
-        var str = this.name + ': ' + this.message;
+        var str = toString(this);
 
         if (defined(this.stack)) {
             str += '\n' + this.stack.toString();
@@ -62,6 +65,10 @@ define([
 
         return str;
     };
+
+    function toString(developerError) {
+        return developerError.name + ': ' + developerError.message;
+    }
 
     /**
      * @private